Fix all compiler warnings (493 → 0) (#3)
Resolve all compiler warnings when building with -Wall -Wextra:

- Fix parentheses warnings (~375): Added parentheses to clarify operator
  precedence in bitwise operations across qrcode and decoder modules
- Fix sign-compare warnings (~62): Added appropriate casts to resolve
  signed/unsigned integer comparison issues
- Fix unused parameter warnings (~29): Marked unused parameters with
  (void) casts in stub implementations
- Fix override-init warnings (18): Added pragma to suppress intentional
  array initializer overwrites in symbol hash table
- Fix fall-through warnings (2): Added fallthrough attributes to
  intentional switch case fall-throughs

All changes preserve existing functionality and maintain backward
compatibility. Build now completes with zero warnings and all tests pass.

+++ zbar/decoder.c
@@ -503,11 +503,11 @@ static char *decoder_dump = NULL;
 static unsigned decoder_dumplen = 0;
 
 const char *_zbar_decoder_buf_dump(unsigned char *buf, unsigned int buflen) {
-  int dumplen = (buflen * 3) + 12;
+  unsigned int dumplen = (buflen * 3) + 12;
   char *p;
-  int i;
+  unsigned int i;
 
-  if (!decoder_dump || dumplen > decoder_dumplen) {
+  if (!decoder_dump || dumplen > (unsigned int)decoder_dumplen) {
     if (decoder_dump)
       free(decoder_dump);
     decoder_dump = malloc(dumplen);
